+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 `TimeZoneLocate` adheres to [Semantic Versioning](http://semver.org/).
 
 ---
+## 0.4.0
+* Support for Swift 4.0
+* Adds extension method ` timeZone(completion:TimeZoneLocateResult)` that uses
+reverse geocoding to fetch a more accurate time zone when connected to the network.
+
 ## 0.3.1
 * Fixes #5: issue relating to empty country code.